week 5/requirements.txt
openai-whisper
pyannote.audio
sounddevice
numpy
transformers
soundfile
torch